refactor: extract copy tree func
